Typed queries for aws-cloudwatch-log-retention integration
Added custom gql definition for `aws-cloudwatch-log-retention` to replace the non-typed `get_aws_accounts` query from `queries.py`. The only option necessary for this integration is `cleanup`, and so that is what the custom gql definition includes.

Modified the the `integration.py` to use the new query, as well as convert its output to a dict in select places to be compatible with other necessary functions.